Conflicts:
	ext/puppetstoredconfigclean.rb
	lib/puppet/defaults.rb
	lib/puppet/rails.rb
	spec/spec_helper.rb
	spec/unit/rails.rb

+++ lib/puppet/rails/database/002_remove_duplicated_index_on_all_tables.rb
@@ -1,17 +1,13 @@
 class RemoveDuplicatedIndexOnAllTables < ActiveRecord::Migration
-    def self.up
-        ActiveRecord::Base.connection.tables.each do |t|
-            if ActiveRecord::Base.connection.indexes(t).collect {|c| c.columns}.include?("id")
-                remove_index t.to_s, :id
-            end
-        end
+  def self.up
+    ActiveRecord::Base.connection.tables.each do |t|
+      remove_index t.to_s, :id if ActiveRecord::Base.connection.indexes(t).collect {|c| c.columns}.include?("id")
     end
+  end
 
-    def self.down
-        ActiveRecord::Base.connection.tables.each do |t|
-            unless ActiveRecord::Base.connection.indexes(t).collect {|c| c.columns}.include?("id")
-                add_index t.to_s, :id, :integer => true
-            end
-        end
+  def self.down
+    ActiveRecord::Base.connection.tables.each do |t|
+      add_index t.to_s, :id, :integer => true unless ActiveRecord::Base.connection.indexes(t).collect {|c| c.columns}.include?("id")
     end
+  end
 end
